+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All Notable changes to `ElementFinder` will be documented in this file
  - #57 Deprecate method `ElementFinder::node` use `ElementFinder::element` instead 
 
 ### Removed
+ - #75 Method `ElementFinder::node`
  - #53 Remove `ArrayAccess` interface from the `StringCollection`, `ObjectCollection` and `ElementCollection`
  - #52 Replace `Iterator`  with `IteratorAggregate` interface inside `StringCollection`, `ObjectCollection` and `ElementCollection`
  - #55 Remove (`StringCollection::prepend`,`StringCollection::addAfter`,`StringCollection::slice`,`StringCollection::extractItems`,`StringCollection::getNext`,`StringCollection::getPrevious`, `StringCollection::append`, `StringCollection::setItems`)